+++ Src/UDateExtractor.pas
@@ -18,7 +18,7 @@ interface
 type
 
   ///  <summary>Method only record that exposes a method that gets either the
-  ///  creation or last-modified date from a file.</summary>
+  ///  creation, last-accessed or last-modified date from a file.</summary>
   TDateExtractor = record
   public
     type
@@ -26,16 +26,27 @@ type
       ///  <summary>Type of date to be extracted from a file.</summary>
       ///  <remarks>
       ///  <para>- LastModified - get date file was last modified.</para>
-      ///  <para>- Created - get date file was created.</para>
+      ///  <para>- Created - get date file was created (Windows only).</para>
+      ///  <para>- StatusChanged - get date when the file's status was was
+      ///  changed (Linux only).</para>
+      ///  <para>- LastAccessed - get date file was last accessed.</para>
       ///  </remarks>
-      TDateType = (LastModified, Created);
+      TDateType = (
+        LastModified,
+        {$IF Defined(MSWINDOWS)}
+        Created,
+        {$ELSEIF Defined(LINUX)}
+        StatusChanged,
+        {$ENDIF}
+        LastAccessed
+      );
       {$SCOPEDENUMS OFF}
   public
-    ///  <summary>Gets either the creation or last-modified date from a file.
-    ///  </summary>
+    ///  <summary>Gets either creation, last-modified or last-accessed date from
+    ///  a file.</summary>
     ///  <param name="FileName">[in] Name of the file to be examined.</param>
-    ///  <param name="DateType">[in] Specifies whether the last-modified or
-    ///  creation date is to be returned.</param>
+    ///  <param name="DateType">[in] Specifies whether the last-modified,
+    ///  last-accessed or creation date is to be returned.</param>
     ///  <returns><c>TDateTime</c>. The required file date.</returns>
     class function GetDate(const FileName: string; const DateType: TDateType):
       TDateTime; static;
@@ -71,8 +82,17 @@ begin
   case DateType of
     TDateType.LastModified:
       Result := TFile.GetLastWriteTime(FileName);
+    {$IF Defined(MSWINDOWS)}
     TDateType.Created:
       Result := TFile.GetCreationTime(FileName);
+    {$ELSEIF Defined(LINUX)}
+    TDateType.StatusChanged:
+      // On Linux, TFile.GetCreationTime ultimately returns the _status.st_ctime
+      // value, which is the status change date. Confusing!
+      Result := TFile.GetCreationTime(FileName);
+    {$ENDIF}
+    TDateType.LastAccessed:
+      Result := TFile.GetLastAccessTime(FileName);
   else
     raise Exception.Create('Invalid TDateExtractor.TDateType value');
   end;